不要なインターフェイスを削除
