+++ src/Beutl.ProjectSystem/NodeTree/IInputSocketForSetter.cs
@@ -1,0 +1,8 @@
+﻿namespace Beutl.NodeTree;
+
+internal interface IInputSocketForSetter : IInputSocket
+{
+    new int LocalId { get; set; }
+
+    void SetProperty(object property);
+}
