--- src/Beutl.ProjectSystem/NodeTree/INodeProperty.cs
@@ -1,5 +1,0 @@
-﻿namespace Beutl.NodeTree;
-
-public interface INodeProperty : INodeItem
-{
-}
