+++ src/Beutl.ProjectSystem/NodeTree/ISocketsCanBeAdded.cs
@@ -7,5 +7,5 @@ public interface ISocketsCanBeAdded
 {
     SocketLocation PossibleLocation { get; }
 
-    bool AddSocket(ISocket socket, [NotNullWhen(true)] out IConnection? connection);
+    bool AddSocket(ISocket socket, [NotNullWhen(true)] out Connection? connection);
 }
