+++ src/Beutl.ProjectSystem/NodeTree/Nodes/Group/IGroupSocket.cs
@@ -1,0 +1,6 @@
+﻿namespace Beutl.NodeTree.Nodes.Group;
+
+public interface IGroupSocket : ISocket
+{
+    CoreProperty? AssociatedProperty { get; set; }
+}
